move custom course dialog component to separate file
